corrected interface prints

+++ pydrone/client_game_pad.py
@@ -31,10 +31,10 @@ print('Connected to ' + joystick_name)
 
 while True:
 
-    axis0 = round(joystick.get_axis(0))
-    axis1 = round(joystick.get_axis(1))
-    lt = round(joystick.get_axis(2))
-    rt = round(joystick.get_axis(5))
+    axis0 = int(round(joystick.get_axis(0)))
+    axis1 = int(round(joystick.get_axis(1)))
+    lt = int(round(joystick.get_axis(2)))
+    rt = int(round(joystick.get_axis(5)))
 
     a = axis0
     b = axis1
